+++ R/inflate-utils.R
@@ -1,3 +1,21 @@
+# There is a function (or R6Class)
+regex_isfunction <- paste(
+  # function
+  "function(\\s*)\\(",
+  # R6Class
+  "R6Class(\\s*)\\(",
+  sep = "|"
+)
+
+regex_extract_fun_name <- paste(
+  # function
+  "[\\w[.]]*(?=(\\s*)(<-|=)(\\s*)function)",
+  # R6Class
+  "[\\w[.]]*(?=(\\s*)(<-|=)(\\s*)R6Class)",
+  # R6::R6Class
+  "[\\w[.]]*(?=(\\s*)(<-|=)(\\s*)R6::R6Class)",
+  sep = "|"
+)
 
 #' Parse function code as tibble and get positions
 #' @param x One row out of function parsed tibble
@@ -6,29 +24,6 @@
 parse_fun <- function(x) { # x <- rmd_fun[3,]
 
   code <- unlist(rmd_node_code(x[["ast"]]))
-  # There is a function (or R6Class)
-  regex_isfunction <- paste(
-    # function
-    "function(\\s*)\\(",
-    # R6Class
-    "R6Class(\\s*)\\(",
-    sep = "|"
-  )
-
-  regex_extract_fun_name <- paste(
-    # function
-    "[\\w[.]]*(?=(\\s*)(<-|=)(\\s*)function)",
-    # R6Class
-    "[\\w[.]]*(?=(\\s*)(<-|=)(\\s*)R6Class)",
-    # R6::R6Class
-    "[\\w[.]]*(?=(\\s*)(<-|=)(\\s*)R6::R6Class)",
-    sep = "|"
-  )
-
-  # stringi::stri_extract_first_regex(
-  #   c("zaza <- function()", "zozo <- R6Class()", "zuzu <- R6::R6Class()"),
-  #   regex_extract_fun_name
-  # )
 
   # Clean extra space between #' and @
   code <- gsub(pattern = "#'\\s*@", "#' @", code)
@@ -36,21 +31,37 @@ parse_fun <- function(x) { # x <- rmd_fun[3,]
   # Get all #'
   all_hastags <- grep("^#'", code)
 
-  # Get all #
-  all_comments <- grep("^#", code)
+  # Get all lines starting with #, including hastags
+  all_comments <- grep("^\\s*#", code)
 
   # Get all functions
   fun_positions <- setdiff(grep(regex_isfunction, code), all_comments)
 
+  # Get lines before "function" if code on multiple lines
+  # Parse only code and not all the rest
+  code_clean_first_fun <- gsub("\\n", " ",
+    as.character(parse(text = code))
+  )
+  code_clean_first_fun <- code_clean_first_fun[grepl(regex_isfunction, code_clean_first_fun)][1]
+
   # find function name
   fun_name <- stringi::stri_extract_first_regex(
-    code[fun_positions],
+    # code[fun_positions],
+    code_clean_first_fun,
     regex_extract_fun_name
   ) %>%
     gsub(" ", "", .) # remove spaces
 
+  if (!is.na(fun_name)) {
+    # Find fun name positions
+    fun_name_positions <- setdiff(grep(fun_name, code, fixed = TRUE), all_comments)
+    # Get the one above 'function()'
+    fun_name_position <- max(fun_name_positions[fun_name_positions <= min(fun_positions)])
+  } else {
+    fun_name_position <- NA
+  }
   # Find start of function
-  first_function_start <- fun_positions[1]
+  first_function_start <- fun_name_position[1]
 
   # Get last hastags above first fun
   if (length(all_hastags) != 0) {
@@ -59,34 +70,48 @@ parse_fun <- function(x) { # x <- rmd_fun[3,]
     last_hastags_above_first_fun <- NA
   }
 
+
+  # TODO deal with empty chunk or non function chunks like datasets
   # Add @noRd if no roxygen doc or no @export before function
-  if (!any(grepl("@export|@noRd", code))) {
+  if (all(grepl("^\\s*$", code))) {
+    # If chunk all empty
+    code <- character(0)
+  } else if (!is.na(first_function_start) &&
+             !any(grepl("@export|@noRd", code[1:first_function_start]))) {
     if (!is.na(last_hastags_above_first_fun)) {
       code <- c(
         code[1:last_hastags_above_first_fun],
         "#' @noRd",
         code[(last_hastags_above_first_fun + 1):length(code)]
       )
-    } else if (all(grepl("^\\s*$", code))) {
-      # If all empty
-      code <- character(0)
-    } else if (!is.na(first_function_start)) {
-      # If there is a function inside
+      fun_name_position <- fun_name_position + 1
+      last_hastags_above_first_fun <- last_hastags_above_first_fun + 1
+    # } else if (all(grepl("^\\s*$", code))) {
+    #   # If all empty
+    #   code <- character(0)
+    } else {
+      # If there is only a function inside
       code <- c("#' @noRd", code)
+      fun_name_position <- fun_name_position + 1
+      last_hastags_above_first_fun <- 1
     }
     # otherwise code stays code
+    # For data documentation for instance
   }
 
   all_arobase <- grep("^#'\\s*@|function(\\s*)\\(", code)
   example_pos_start <- grep("^#'\\s*@example", code)[1]
 
   example_pos_end <- all_arobase[all_arobase > example_pos_start][1] - 1
-  example_pos_end <- ifelse(is.na(example_pos_end),
-    grep("function(\\s*)\\(", code) - 1,
+  example_pos_end <- ifelse(
+    is.na(example_pos_end),
+    last_hastags_above_first_fun,
+    # fun_name_position - 1,
+    #grep("function(\\s*)\\(", code) - 1,
     example_pos_end
   )
 
-  # Get @rdname and @filename for groups
+  # Get @rdname and @filename for grouping functions
   tag_filename <- gsub(
     "^#'\\s*@filename\\s*", "",
     code[grep("^#'\\s*@filename", code)]
@@ -117,7 +142,7 @@ parse_fun <- function(x) { # x <- rmd_fun[3,]
 add_names_to_parsed <- function(parsed_tbl, fun_code) {
   # Which parts were functions
   which_parsed_fun <- which(!is.na(parsed_tbl$label) &
-    grepl(regex_functions, parsed_tbl$label))
+                              grepl(regex_functions, parsed_tbl$label))
 
   # From fun_code, we retrieve fun_name & rox_filename
   parsed_tbl[["fun_name"]] <- NA_character_
@@ -150,7 +175,7 @@ add_names_to_parsed <- function(parsed_tbl, fun_code) {
         sec_title_name[["sec_title"]] == x, "sec_fun_name"
       ]
       parsed_tbl[group, "sec_fun_name"] <- ifelse(length(sec_fun_name) == 0,
-        NA_character_, as.character(sec_fun_name)
+                                                  NA_character_, as.character(sec_fun_name)
       )
       parsed_tbl[group, ] <- tidyr::fill(
         parsed_tbl[group, ],
@@ -179,7 +204,7 @@ add_names_to_parsed <- function(parsed_tbl, fun_code) {
   pkg_filled[["file_name"]] <- NA_character_
   # chunk_filename
   pkg_filled[["file_name"]] <- ifelse(!is.na(pkg_filled[["chunk_filename"]]),
-    pkg_filled[["chunk_filename"]], NA_character_
+                                      pkg_filled[["chunk_filename"]], NA_character_
   )
   # rox_filename
   pkg_filled[["file_name"]] <- ifelse(
@@ -255,7 +280,7 @@ parse_test <- function(x, pkg, relative_flat_file) { # x <- rmd_test[1,]
 add_fun_code_examples <- function(parsed_tbl, fun_code) {
   # Example in separate chunk
   which_parsed_ex <- which(!is.na(parsed_tbl$label) &
-    grepl(regex_example, parsed_tbl$label))
+                             grepl(regex_example, parsed_tbl$label))
   rmd_ex <- parsed_tbl[which_parsed_ex, ]
 
   # Get file_name variable
@@ -267,8 +292,8 @@ add_fun_code_examples <- function(parsed_tbl, fun_code) {
   fun_code <- fun_code[order(fun_code[["order"]]), ]
   # Get file_name for not functions. Only last place where possible
   fun_code[["file_name"]] <- ifelse(is.na(fun_code[["file_name"]]),
-    fun_code[["sec_title"]],
-    fun_code[["file_name"]]
+                                    fun_code[["sec_title"]],
+                                    fun_code[["file_name"]]
   )
 
   #  Example already in skeleton
@@ -353,8 +378,8 @@ add_fun_code_examples <- function(parsed_tbl, fun_code) {
     }
 
     end_skeleton <- ifelse(is.na(fun_code_x[["example_pos_start"]]),
-      fun_code_x[["example_pos_end"]],
-      fun_code_x[["example_pos_start"]] - 1
+                           fun_code_x[["example_pos_end"]],
+                           fun_code_x[["example_pos_start"]] - 1
     )
 
     all_fun_code <- stats::na.omit(c(
@@ -447,14 +472,14 @@ create_vignette_head <- function(pkg, vignette_name, yaml_options = NULL) {
 title: ".{vignette_title}."
 output: rmarkdown::html_vignette',
       ifelse(length(yaml_options) != 0,
-        glue::glue_collapse(
-          c(
-            "",
-            glue("{names(yaml_options)}: \"{yaml_options}\""), ""
-          ),
-          sep = "\n"
-        ),
-        "\n"
+             glue::glue_collapse(
+               c(
+                 "",
+                 glue("{names(yaml_options)}: \"{yaml_options}\""), ""
+               ),
+               sep = "\n"
+             ),
+             "\n"
       ),
       'vignette: >
   %\\VignetteIndexEntry{.{asciify_name(vignette_name)}.}
@@ -473,7 +498,7 @@ knitr::opts_chunk$set(
 library(.{pkgname}.)
 ```
     ',
-      .open = ".{", .close = "}."
+    .open = ".{", .close = "}."
     )
   )
 }