+++ NEWS.md
@@ -7,6 +7,7 @@
 ## Bug fixes
 
 - Fix using `function(` in documentation (#174, @FlorenceMounier)
+- Fix using line break after function name in flat files (#142, @FlorenceMounier)
 
 # fusen 0.4.1